REF: remove dead code and add more structure
